Move DirectRunner specific classes into direct runner package.
